refactor: move constants to const;isort
