+++ hq_superset/tests/test_oauth.py
@@ -5,11 +5,8 @@ from flask import session
 
 from hq_superset.exceptions import OAuthSessionExpired
 from hq_superset.oauth import get_valid_cchq_oauth_token
-from hq_superset.utils import (
-    SESSION_OAUTH_RESPONSE_KEY,
-    SESSION_USER_DOMAINS_KEY,
-)
 
+from ..const import SESSION_OAUTH_RESPONSE_KEY, SESSION_USER_DOMAINS_KEY
 from .base_test import SupersetTestCase
